lead-analysis/requirements.txt
fastapi
uvicorn
beautifulsoup4
python-dotenv
pydantic
langchain
langchain-huggingface
langchain-openai